match names and versions with validating regex
this is probably a terrible idea from a performance perspective.
but while we're talking about that, I'm pretty sure instantiating
all of these objects is pretty terrible from a performance
perspective, too. I think maybe we need something that can reach
into the deps hash and returns specs on command instead of actually
providing an array of every spec.